requirements.txt
streamlit
python-dotenv
requests
fuzzywuzzy
python-Levenshtein





